fix(test case): fix test case for directory tree component
